requirements.txt
atlassian-python-api
PyGithub
gitpython
pytz
python-dotenv